+++ Mindspace_back/src/initData/neo4jNodeData.ts
@@ -1,0 +1,154 @@
+export const nodeNames = [
+  'Javascript',
+  'Webpack',
+  'React',
+  'JavaScript Immutability',
+  'Ajax',
+  'React Route',
+  'State Management',
+  'Class vs Component',
+  'React Query',
+  'Jest',
+  'ES6',
+  'Redux',
+  'Mobx',
+  'Zustand',
+  'Recoil',
+  'Redux saga',
+  'Redux toolkit',
+  'Browser Render',
+  'HTML',
+  'Proxy',
+  'Auth',
+  'OAuth2.0',
+  'FaceBook Login',
+  'Kakao Login',
+  'Bearer Auth',
+  'JWT Token',
+  'CSS',
+  'CSS vs CSS in JS',
+  'SASS',
+  'Tailwind',
+  'Bootstrap',
+  'Styled-compoent',
+  'Material-UI',
+];
+
+export const nodeLinks = [
+  {
+    source: 1,
+    target: 2,
+  },
+  {
+    source: 1,
+    target: 4,
+  },
+  {
+    source: 1,
+    target: 5,
+  },
+  {
+    source: 3,
+    target: 6,
+  },
+  {
+    source: 3,
+    target: 7,
+  },
+  {
+    source: 3,
+    target: 8,
+  },
+  {
+    source: 3,
+    target: 9,
+  },
+  {
+    source: 3,
+    target: 10,
+  },
+  {
+    source: 1,
+    target: 11,
+  },
+  {
+    source: 7,
+    target: 12,
+  },
+  {
+    source: 7,
+    target: 13,
+  },
+  {
+    source: 7,
+    target: 14,
+  },
+  {
+    source: 7,
+    target: 15,
+  },
+  {
+    source: 12,
+    target: 16,
+  },
+  {
+    source: 12,
+    target: 17,
+  },
+  {
+    source: 18,
+    target: 19,
+  },
+  {
+    source: 5,
+    target: 20,
+  },
+  {
+    source: 21,
+    target: 22,
+  },
+  {
+    source: 21,
+    target: 25,
+  },
+  {
+    source: 21,
+    target: 26,
+  },
+  {
+    source: 22,
+    target: 23,
+  },
+  {
+    source: 22,
+    target: 24,
+  },
+  {
+    source: 27,
+    target: 19,
+  },
+  {
+    source: 27,
+    target: 28,
+  },
+  {
+    source: 27,
+    target: 29,
+  },
+  {
+    source: 27,
+    target: 30,
+  },
+  {
+    source: 27,
+    target: 31,
+  },
+  {
+    source: 27,
+    target: 32,
+  },
+  {
+    source: 27,
+    target: 33,
+  },
+];
