+++ Mindspace_back/src/initData/nodeData.ts
@@ -1,0 +1,35 @@
+export const initialData = [
+  { id: 1, name: 'Javascript' },
+  { id: 2, name: 'Webpack' },
+  { id: 3, name: 'React' },
+  { id: 4, name: 'JavaScript Immutability' },
+  { id: 5, name: 'Ajax' },
+  { id: 6, name: 'React Route' },
+  { id: 7, name: 'State Management' },
+  { id: 8, name: 'Class vs Component' },
+  { id: 9, name: 'React Query' },
+  { id: 10, name: 'Jest' },
+  { id: 11, name: 'ES6' },
+  { id: 12, name: 'Redux' },
+  { id: 13, name: 'Mobx' },
+  { id: 14, name: 'Zustand' },
+  { id: 15, name: 'Recoil' },
+  { id: 16, name: 'Redux saga' },
+  { id: 17, name: 'Redux toolkit' },
+  { id: 18, name: 'Browser Render' },
+  { id: 19, name: 'HTML' },
+  { id: 20, name: 'Proxy' },
+  { id: 21, name: 'Auth' },
+  { id: 22, name: 'OAuth2.0' },
+  { id: 23, name: 'FaceBook Login' },
+  { id: 24, name: 'Kakao Login' },
+  { id: 25, name: 'Bearer Auth' },
+  { id: 26, name: 'JWT Token' },
+  { id: 27, name: 'CSS' },
+  { id: 28, name: 'CSS vs CSS in JS' },
+  { id: 29, name: 'SASS' },
+  { id: 30, name: 'Tailwind' },
+  { id: 31, name: 'Bootstrap' },
+  { id: 32, name: 'Styled-compoent' },
+  { id: 33, name: 'Material-UI' },
+];
