+++ Mindspace_back/src/node/repository/node.repository.ts
@@ -1,0 +1,24 @@
+import { InjectRepository } from '@nestjs/typeorm';
+import { Node } from '../entities/node.entity';
+import { Repository } from 'typeorm';
+
+export class CustomNodeRepository {
+  constructor(
+    @InjectRepository(Node)
+    private readonly nodeRepository: Repository<Node>,
+  ) {}
+
+  async isNodeWrittenByUser(
+    nodeName: string,
+    userId: number,
+  ): Promise<boolean> {
+    const count = await this.nodeRepository
+      .createQueryBuilder('node')
+      .innerJoin('node.boards', 'board')
+      .where('node.name = :nodeName', { nodeName })
+      .andWhere('board.userId = :userId', { userId })
+      .getCount();
+
+    return count > 0;
+  }
+}
